feat: 출고 정보 업데이트 시 부가서비스 저장 로직의 변경 #18567

See merge request onchannel/1688!327

+++ app/Services/Wms/WmsW1.php
@@ -699,9 +699,11 @@ class WmsW1 extends WmsAbstract
                         }
                     }
 
-                    BonaeraOutDeliveryExtraData::where("group_no", $groupNo)
-                    ->whereNotIn("extra_name", $existsExtras)
-                    ->forceDelete();
+                    if( !empty($existsExtras)) {
+                        BonaeraOutDeliveryExtraData::where("group_no", $groupNo)
+                        ->whereNotIn("extra_name", $existsExtras)
+                        ->forceDelete();
+                    }
                 }
 
                 if( isset($res["weightList"][0]) ){
@@ -766,9 +768,11 @@ class WmsW1 extends WmsAbstract
                         }
                     }
 
-                    BonaeraOutExtraData::where("sh_no", $shNo)
-                    ->whereNotIn("extra_name", $existsExtras)
-                    ->forceDelete();
+                    if( !empty($existsExtras)) {
+                        BonaeraOutExtraData::where("sh_no", $shNo)
+                        ->whereNotIn("extra_name", $existsExtras)
+                        ->forceDelete();
+                    }
                 }
             }
         } catch (Exception $e) {
